Refactor proof-setup to add a level of indirection
This allows us to have a "single" module, and an "all" module, with the
latter containing logic for doing all of the setups at once, and
potentially in parallel.
