+++ crates/crypto/proof-setup/src/single/group.rs
